This hooks up the new RANKS placeholders.
